Revert "fix(PeriphDrivers): Fix `-wstrict-prototype` and other Misc. MAX78002 Warnings (#1077)"
This reverts commit fca56e9cfb1d25169aa64e37c9296d16cfb5b5dd.

+++ Examples/MAX78002/QSPI/main.c
@@ -236,7 +236,7 @@ int main(void)
 
     if (fail_count > 0) {
         printf("\nFailed with %i mismatches (%.2f%%)!\n", fail_count,
-               (double)(100 * (((float)fail_count) / (TEST_SIZE * TEST_COUNT))));
+               100 * (((float)fail_count) / (TEST_SIZE * TEST_COUNT)));
         return E_FAIL;
     }
